+++ src/snowflake/cli/api/project/schemas/native_app/application.py
@@ -21,6 +21,7 @@ from snowflake.cli.api.project.schemas.updatable_model import (
     IdentifierField,
     UpdatableModel,
 )
+from snowflake.cli.api.project.util import get_sanitized_username
 
 
 class SqlScriptHookType(UpdatableModel):
@@ -51,4 +52,12 @@ class Application(UpdatableModel):
     post_deploy: Optional[List[ApplicationPostDeployHook]] = Field(
         title="Actions that will be executed after the application object is created/upgraded",
         default=None,
+    )
+
+
+class ApplicationV11(Application):
+    # Templated defaults only supported in v1.1+
+    name: Optional[str] = Field(
+        title="Name of the application object created when you run the snow app run command",
+        default=f"<% id_concat(ctx.native_app.name, '_{get_sanitized_username()}') %>",
     )