+++ src/snowflake/cli/api/project/schemas/native_app/path_mapping.py
@@ -16,8 +16,11 @@ from __future__ import annotations
 
 from typing import Any, Dict, List, Optional, Union
 
-from pydantic import Field, field_validator
-from snowflake.cli.api.project.schemas.updatable_model import UpdatableModel
+from pydantic import Field
+from snowflake.cli.api.project.schemas.updatable_model import (
+    UpdatableModel,
+    field_validator_allowing_templates,
+)
 
 
 class ProcessorMapping(UpdatableModel):
@@ -46,10 +49,10 @@ class PathMapping(UpdatableModel):
         default=[],
     )
 
-    @field_validator("processors")
-    @classmethod
+    @field_validator_allowing_templates("processors")
     def transform_processors(
-        cls, input_values: Optional[List[Union[str, Dict, ProcessorMapping]]]
+        cls,  # noqa: N805, classmethod included
+        input_values: Optional[List[Union[str, Dict, ProcessorMapping]]],
     ) -> List[ProcessorMapping]:
         if input_values is None:
             return []