+++ src/snowflake/cli/api/project/schemas/updatable_model.py
@@ -14,17 +14,125 @@
 
 from __future__ import annotations
 
-from typing import Any, Dict
+from contextlib import contextmanager
+from contextvars import ContextVar
+from typing import Any, Dict, Iterator, Optional
 
-from pydantic import BaseModel, ConfigDict, Field
+from pydantic import (
+    BaseModel,
+    ConfigDict,
+    Field,
+    ValidationInfo,
+    field_validator,
+)
+from pydantic.fields import FieldInfo
 from snowflake.cli.api.project.util import IDENTIFIER_NO_LENGTH
+
+PROJECT_TEMPLATE_START = "<%"
+
+
+def _is_templated(info: ValidationInfo, value: Any) -> bool:
+    return (
+        info.context
+        and info.context.get("includes_templates", False)
+        and isinstance(value, str)
+        and PROJECT_TEMPLATE_START in value
+    )
+
+
+def field_validator_allowing_templates(*validator_args, **validator_kwargs):
+    """
+    This validator replaces field_validator from Pydantic.
+    The difference is that this validator will skip validation
+    whenever a value is a String and the value is templated.
+
+    It also checks the context to ensure includes_templates is set to True.
+    Otherwise, if includes_templates context is not set, it behaves the same
+    way as field_validator, and does not check templates anymore.
+    """
+    mode = validator_kwargs.get("mode", "after")
+
+    def decorator(func):
+        def wrapper_mode_wrap(cls, value, handler, info: ValidationInfo, **kwargs):
+            if _is_templated(info, value):
+                return value
+            return func(cls, value, handler, **kwargs)
+
+        def wrapper_default(cls, value, info: ValidationInfo, **kwargs):
+            if _is_templated(info, value):
+                return value
+            return func(cls, value, **kwargs)
+
+        wrapper = wrapper_mode_wrap if mode == "wrap" else wrapper_default
+        return field_validator(*validator_args, **validator_kwargs)(
+            classmethod(wrapper)
+        )
+
+    return decorator
+
+
+_initial_context: ContextVar[Optional[Dict[str, Any]]] = ContextVar(
+    "_init_context_var", default=None
+)
+
+
+@contextmanager
+def context(value: Dict[str, Any]) -> Iterator[None]:
+    """
+    Thread safe context for Pydantic.
+    By using `with context()`, you ensure context changes apply
+    to the with block only
+    """
+    token = _initial_context.set(value)
+    try:
+        yield
+    finally:
+        _initial_context.reset(token)
 
 
 class UpdatableModel(BaseModel):
     model_config = ConfigDict(validate_assignment=True, extra="forbid")
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(**kwargs)
+    def __init__(self, /, **data: Any) -> None:
+        self.__pydantic_validator__.validate_python(
+            data,
+            self_instance=self,
+            context=_initial_context.get(),
+        )
+
+    @classmethod
+    def __init_subclass__(cls, **kwargs):
+        super().__init_subclass__(**kwargs)
+
+        # Collect all the Pydantic fields
+        # from all the classes in the inheritance chain
+        all_fields = {}
+        for class_ in cls.__mro__:
+            if hasattr(class_, "__annotations__"):
+                all_fields.update(class_.__annotations__)
+
+        class_dict = cls.__dict__
+        for field in all_fields:
+            if field in class_dict and isinstance(class_dict[field], FieldInfo):
+                # exclude EntityTypeField fields (which are instantiated with is_type_field = true)
+                schema_extra = class_dict[field].json_schema_extra
+                if (
+                    not schema_extra
+                    or "is_type_field" not in schema_extra
+                    or not schema_extra["is_type_field"]
+                ):
+                    cls._add_validator(field)
+
+    @classmethod
+    def _add_validator(cls, field_name: str):
+        def _validator(cls, value, handler):
+            return handler(value)
+
+        setattr(
+            cls,
+            f"_template_validate_{field_name}",
+            field_validator_allowing_templates(field_name, mode="wrap")(_validator),
+        )
 
     def update_from_dict(self, update_values: Dict[str, Any]):
         """
@@ -47,5 +155,9 @@ class UpdatableModel(BaseModel):
         return self
 
 
-def IdentifierField(*args, **kwargs):  # noqa
+def EntityTypeField(*args, **kwargs):  # noqa N802
+    return Field(is_type_field=True, *args, **kwargs)
+
+
+def IdentifierField(*args, **kwargs):  # noqa N802
     return Field(max_length=254, pattern=IDENTIFIER_NO_LENGTH, *args, **kwargs)